MVP finalized with demonstration
This program works well "out of the box" but is still missing some important functionality:
- headers
- missing values to NULL
- support for data types other than int, char, double, and string
- support for delimiters in strings and double quote enclosures

I also need to do a more careful scan over to check for any memory leaks because there are a lot of finicky void pointers and pointers to void pointers. I also need to implement several unit tests.