removed null char assertions
I realized that with a single column CSV, the best/most intuitive separator to use is the null char (unless you want to program in another argument to csv_create(), which I don't). Therefore, two assertions in the private functions needed to be removed to let this happen. (I am running this program without -DNDEBUG for extra safety checks until I've swept for memory leaks).