Fixed memory leak with double pointer
Regarding the last commit, turns out I was being an idiot and DynamoRio did flag a correct issue. darray_append input pointer would become invalid when realloc was called as the structure could move in memory but the input pointer did not change. So not only was a massive amount of memory leaking, but the darray was pointing to an invalid location on most calls, getting worse as the array grows. 

The fix of course was simple, pass a pointer to darray on darray_append so that we can modify its value after realloc. Isolating the issue took 4 or 5 hours...