added queue pop functionality
I FINALLY found a use for the extra 8 reserved bytes that I had in the array header, which I only really had in the first place to keep the structure nicely packed for variably sized array items.

Now the extra 8 bytes is used for the popleft function. Since a queue implementation on a dynamic array involves pushing the entire array block of memory back by one index, we can't just return a pointer to the popleft item since it will disappear. This is annoying since the stack pop and peek functions do return pointers, and return NULL for ease of use when the array is empty. I wanted the same pointer functionality for the queue popleft operation. 

So, that 8 byte reserved section in the header is now used for the queue. An item is popped from the front, but copied to the queue cache. This queue cache pointer gets returned to the user, and we can safely move back the items in the array and overwrite the previous data, since the cache has all the info that the user needs.