refactored to generic ADT
thanks to the typedef'd array_item, void* items are basically a free bonus with the current implementation. I just needed to add some documentation on the pop functions to warn the client of popping with NULL arguments, which may cause a memory leak.

I also removed the darray_show() function, which I had mostly used for debugging. It clashes too much with the generic ADT, a custom printf by the client directly on the exposed array would be the better choice.