API change for backend preference
The previous iteration was friendly for end-user needs, but I realized I'll never use this API for anything other than backend programmer implementations. So, I removed the error codes on the pop/peek functions and swapped them out for assertions. I also added assertions onto the other functions.

Secondly, I did not like the API functionality for having to pass a parameter to receive popped items. Especially for generics, the (void**) my_var call is a nuisance. Wit h the new focus on backend, I refactored the functions to directly return the popped array_items. To do this I modified the struct header to add a queue cache and a secondary unused cache (to keep the packing consistent).

In summary, I think the new API is much, much easier to use. And the implementation is better oriented to fail early and fail close to the error.